backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
langchain
langchain-core
langchain-openai
qdrant-client
psycopg2-binary
PyMuPDF
SQLAlchemy
alembic
python-dotenv
pytest
httpx
tiktoken
lxml
python-multipart
